% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiverse_project-package.R
\docType{package}
\name{epiverse_project-package}
\alias{epiverse_project}
\alias{epiverse_project-package}
\title{epiverse_project: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eaferguson/epiverse_project}
  \item Report bugs at \url{https://github.com/eaferguson/epiverse_project/issues}
}

}
\author{
\strong{Maintainer}: Elaine Ferguson \email{elaine.ferguson@glasgow.ac.uk} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
